Move categories.js to js folder
