+++ pages/_category/browser/index.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-import categories from '@/assets/categories'
+import categories from '@/assets/js/categories'
 
 export default {
   data() {
